+++ zerver/tests/test_invite.py
@@ -29,7 +29,7 @@ from zerver.actions.create_user import (
     process_new_human_user,
     set_up_streams_for_new_human_user,
 )
-from zerver.actions.default_streams import do_add_default_stream
+from zerver.actions.default_streams import do_add_default_stream, do_remove_default_stream
 from zerver.actions.invites import (
     do_create_multiuse_invite_link,
     do_get_invites_controlled_by_user,
@@ -137,6 +137,7 @@ class StreamSetupTest(ZulipTestCase):
             admin,
             [new_user_email],
             streams,
+            include_realm_default_subscriptions=False,
             invite_expires_in_minutes=1000,
         )
 
@@ -177,6 +178,7 @@ class InviteUserBase(ZulipTestCase):
         invite_expires_in_minutes: Optional[int] = INVITATION_LINK_VALIDITY_MINUTES,
         body: str = "",
         invite_as: int = PreregistrationUser.INVITE_AS["MEMBER"],
+        include_realm_default_subscriptions: bool = False,
         realm: Optional[Realm] = None,
     ) -> "TestHttpResponse":
         """
@@ -201,6 +203,9 @@ class InviteUserBase(ZulipTestCase):
                     "invite_expires_in_minutes": invite_expires_in,
                     "stream_ids": orjson.dumps(stream_ids).decode(),
                     "invite_as": invite_as,
+                    "include_realm_default_subscriptions": orjson.dumps(
+                        include_realm_default_subscriptions
+                    ).decode(),
                 },
                 subdomain=realm.string_id if realm else "zulip",
             )
@@ -795,13 +800,44 @@ class InviteUserTest(InviteUserBase):
         self.assert_json_success(self.invite(invitee, []))
         self.assertTrue(find_key_by_email(invitee))
 
-        default_streams = get_default_streams_for_realm_as_dicts(realm.id)
+        default_streams = get_slim_realm_default_streams(realm.id)
         self.assert_length(default_streams, 3)
 
         self.submit_reg_form_for_user(invitee, "password")
-        # If no streams are provided, user is not subscribed to
-        # default streams as well.
+        # If no streams are provided, user is not subscribed to default
+        # streams as well if include_realm_default_subscriptions is False.
         self.check_user_subscribed_only_to_streams("bob", [])
+
+        verona = get_stream("Verona", realm)
+        sandbox = get_stream("sandbox", realm)
+        zulip = get_stream("Zulip", realm)
+        default_streams = get_slim_realm_default_streams(realm.id)
+        self.assert_length(default_streams, 3)
+        self.assertCountEqual(default_streams, [verona, sandbox, zulip])
+
+        # Check that user is subscribed to the streams that were set as default
+        # at the time of account creation and not at the time of inviting them.
+        invitee = self.nonreg_email("test")
+        self.assert_json_success(self.invite(invitee, [], include_realm_default_subscriptions=True))
+        self.assertTrue(find_key_by_email(invitee))
+
+        denmark = get_stream("Denmark", realm)
+        do_add_default_stream(denmark)
+        do_remove_default_stream(verona)
+        self.submit_reg_form_for_user(invitee, "password")
+        self.check_user_subscribed_only_to_streams("test", [denmark, sandbox, zulip])
+
+        default_streams = get_slim_realm_default_streams(realm.id)
+        self.assertCountEqual(default_streams, [denmark, sandbox, zulip])
+        invitee = self.nonreg_email("test1")
+        self.assert_json_success(
+            self.invite(invitee, [verona.name], include_realm_default_subscriptions=True)
+        )
+        self.assertTrue(find_key_by_email(invitee))
+        # Check that the user is subscribed to both default streams and stream
+        # passed in streams list.
+        self.submit_reg_form_for_user(invitee, "password")
+        self.check_user_subscribed_only_to_streams("test1", [denmark, sandbox, verona, zulip])
 
     def test_can_invite_others_to_realm(self) -> None:
         def validation_func(user_profile: UserProfile) -> bool:
@@ -1303,7 +1339,7 @@ so we didn't send them an invitation. We did send invitations to everyone else!"
 
         # User will be subscribed to default streams even when the
         # referrer does not have permission to subscribe others.
-        result = self.invite(invitee, [])
+        result = self.invite(invitee, [], include_realm_default_subscriptions=True)
         self.assert_json_success(result)
         self.check_sent_emails([invitee])
 
@@ -1312,6 +1348,19 @@ so we didn't send them an invitation. We did send invitations to everyone else!"
         default_streams = get_slim_realm_default_streams(realm.id)
         self.assert_length(default_streams, 3)
         self.check_user_subscribed_only_to_streams("alice", default_streams)
+        mail.outbox.pop()
+
+        invitee = self.nonreg_email("newguy")
+        # The invited user will not be subscribed to default streams as well.
+        result = self.invite(invitee, [], include_realm_default_subscriptions=False)
+        self.assert_json_success(result)
+        self.check_sent_emails([invitee])
+
+        self.submit_reg_form_for_user(invitee, "password")
+        default_streams = get_slim_realm_default_streams(realm.id)
+        self.assert_length(default_streams, 3)
+
+        self.check_user_subscribed_only_to_streams("newguy", [])
         mail.outbox.pop()
 
         self.login("iago")
@@ -1332,14 +1381,14 @@ so we didn't send them an invitation. We did send invitations to everyone else!"
         mail.outbox.pop()
 
         invitee = self.nonreg_email("test1")
-        # User will not be subscribed to any streams, because the streams
-        # list is empty when referrer has permission to subscribe others.
-        result = self.invite(invitee, [])
+        result = self.invite(invitee, [], include_realm_default_subscriptions=True)
         self.assert_json_success(result)
         self.check_sent_emails([invitee])
 
         self.submit_reg_form_for_user(invitee, "password")
-        self.check_user_subscribed_only_to_streams("test1", [])
+        self.check_user_subscribed_only_to_streams(
+            "test1", get_slim_realm_default_streams(realm.id)
+        )
 
     def test_invitation_reminder_email(self) -> None:
         # All users belong to zulip realm
@@ -1512,6 +1561,7 @@ so we didn't send them an invitation. We did send invitations to everyone else!"
                 self.user_profile,
                 ["foo@zulip.com"],
                 streams,
+                include_realm_default_subscriptions=False,
                 invite_expires_in_minutes=invite_expires_in_minutes,
             )
         prereg_user = PreregistrationUser.objects.get(email="foo@zulip.com")
@@ -1520,12 +1570,14 @@ so we didn't send them an invitation. We did send invitations to everyone else!"
                 self.user_profile,
                 ["foo@zulip.com"],
                 streams,
+                include_realm_default_subscriptions=False,
                 invite_expires_in_minutes=invite_expires_in_minutes,
             )
             do_invite_users(
                 self.user_profile,
                 ["foo@zulip.com"],
                 streams,
+                include_realm_default_subscriptions=False,
                 invite_expires_in_minutes=invite_expires_in_minutes,
             )
 
@@ -1537,6 +1589,7 @@ so we didn't send them an invitation. We did send invitations to everyone else!"
                 lear_user,
                 ["foo@zulip.com"],
                 [],
+                include_realm_default_subscriptions=True,
                 invite_expires_in_minutes=invite_expires_in_minutes,
             )
 
@@ -1748,37 +1801,49 @@ class InvitationsTestCase(InviteUserBase):
                 user_profile,
                 ["TestOne@zulip.com"],
                 streams,
+                include_realm_default_subscriptions=False,
                 invite_expires_in_minutes=invite_expires_in_minutes,
             )
             do_invite_users(
                 user_profile,
                 ["TestTwo@zulip.com"],
                 streams,
+                include_realm_default_subscriptions=False,
                 invite_expires_in_minutes=invite_expires_in_minutes,
             )
             do_invite_users(
                 hamlet,
                 ["TestThree@zulip.com"],
                 streams,
+                include_realm_default_subscriptions=False,
                 invite_expires_in_minutes=invite_expires_in_minutes,
             )
             do_invite_users(
                 othello,
                 ["TestFour@zulip.com"],
                 streams,
+                include_realm_default_subscriptions=False,
                 invite_expires_in_minutes=invite_expires_in_minutes,
             )
             do_invite_users(
                 self.mit_user("sipbtest"),
                 ["TestOne@mit.edu"],
                 [],
+                include_realm_default_subscriptions=False,
                 invite_expires_in_minutes=invite_expires_in_minutes,
             )
+
         do_create_multiuse_invite_link(
-            user_profile, PreregistrationUser.INVITE_AS["MEMBER"], invite_expires_in_minutes
+            user_profile,
+            PreregistrationUser.INVITE_AS["MEMBER"],
+            invite_expires_in_minutes,
+            include_realm_default_subscriptions=False,
         )
         do_create_multiuse_invite_link(
-            hamlet, PreregistrationUser.INVITE_AS["MEMBER"], invite_expires_in_minutes
+            hamlet,
+            PreregistrationUser.INVITE_AS["MEMBER"],
+            invite_expires_in_minutes,
+            include_realm_default_subscriptions=False,
         )
         self.assert_length(do_get_invites_controlled_by_user(user_profile), 6)
         self.assert_length(do_get_invites_controlled_by_user(hamlet), 2)
@@ -1808,6 +1873,7 @@ class InvitationsTestCase(InviteUserBase):
                 user_profile,
                 ["TestOne@zulip.com"],
                 streams,
+                include_realm_default_subscriptions=False,
                 invite_expires_in_minutes=invite_expires_in_minutes,
             )
 
@@ -1818,10 +1884,14 @@ class InvitationsTestCase(InviteUserBase):
                 user_profile,
                 ["TestTwo@zulip.com"],
                 streams,
+                include_realm_default_subscriptions=False,
                 invite_expires_in_minutes=invite_expires_in_minutes,
             )
             do_create_multiuse_invite_link(
-                othello, PreregistrationUser.INVITE_AS["MEMBER"], invite_expires_in_minutes
+                othello,
+                PreregistrationUser.INVITE_AS["MEMBER"],
+                invite_expires_in_minutes,
+                include_realm_default_subscriptions=False,
             )
 
         prereg_user_three = PreregistrationUser(
@@ -1835,7 +1905,10 @@ class InvitationsTestCase(InviteUserBase):
         )
 
         do_create_multiuse_invite_link(
-            hamlet, PreregistrationUser.INVITE_AS["MEMBER"], invite_expires_in_minutes
+            hamlet,
+            PreregistrationUser.INVITE_AS["MEMBER"],
+            invite_expires_in_minutes,
+            include_realm_default_subscriptions=False,
         )
 
         result = self.client_get("/json/invites")
@@ -1865,19 +1938,27 @@ class InvitationsTestCase(InviteUserBase):
                 user_profile,
                 ["TestOne@zulip.com"],
                 streams,
+                include_realm_default_subscriptions=False,
                 invite_expires_in_minutes=None,
             )
             do_invite_users(
                 user_profile,
                 ["TestTwo@zulip.com"],
                 streams,
+                include_realm_default_subscriptions=False,
                 invite_expires_in_minutes=100 * 24 * 60,
             )
             do_create_multiuse_invite_link(
-                user_profile, PreregistrationUser.INVITE_AS["MEMBER"], None
+                user_profile,
+                PreregistrationUser.INVITE_AS["MEMBER"],
+                None,
+                include_realm_default_subscriptions=False,
             )
             do_create_multiuse_invite_link(
-                user_profile, PreregistrationUser.INVITE_AS["MEMBER"], 100
+                user_profile,
+                PreregistrationUser.INVITE_AS["MEMBER"],
+                100,
+                include_realm_default_subscriptions=False,
             )
 
         result = self.client_get("/json/invites")
@@ -2311,9 +2392,16 @@ class MultiuseInviteTest(ZulipTestCase):
         self.realm.save()
 
     def generate_multiuse_invite_link(
-        self, streams: Optional[List[Stream]] = None, date_sent: Optional[datetime] = None
+        self,
+        streams: Optional[List[Stream]] = None,
+        date_sent: Optional[datetime] = None,
+        include_realm_default_subscriptions: bool = False,
     ) -> str:
-        invite = MultiuseInvite(realm=self.realm, referred_by=self.example_user("iago"))
+        invite = MultiuseInvite(
+            realm=self.realm,
+            referred_by=self.example_user("iago"),
+            include_realm_default_subscriptions=include_realm_default_subscriptions,
+        )
         invite.save()
 
         if streams is not None:
@@ -2411,9 +2499,13 @@ class MultiuseInviteTest(ZulipTestCase):
         name1 = "newuser"
         name2 = "bob"
         name3 = "alice"
+        name4 = "test"
+        name5 = "test1"
         email1 = self.nonreg_email(name1)
         email2 = self.nonreg_email(name2)
         email3 = self.nonreg_email(name3)
+        email4 = self.nonreg_email(name4)
+        email5 = self.nonreg_email(name5)
 
         stream_names = ["Rome", "Scotland", "Venice"]
         streams = [get_stream(stream_name, self.realm) for stream_name in stream_names]
@@ -2428,11 +2520,32 @@ class MultiuseInviteTest(ZulipTestCase):
         self.check_user_subscribed_only_to_streams(name2, streams)
 
         streams = []
-        invite_link = self.generate_multiuse_invite_link(streams=streams)
+        invite_link = self.generate_multiuse_invite_link(
+            streams=streams, include_realm_default_subscriptions=True
+        )
         self.check_user_able_to_register(email3, invite_link)
-        # User is not subscribed to default streams as well.
-        self.assert_length(get_default_streams_for_realm_as_dicts(self.realm.id), 3)
-        self.check_user_subscribed_only_to_streams(name3, [])
+        self.assert_length(get_slim_realm_default_streams(self.realm.id), 3)
+        self.check_user_subscribed_only_to_streams(
+            name3, get_slim_realm_default_streams(self.realm.id)
+        )
+
+        invite_link = self.generate_multiuse_invite_link(
+            streams=streams, include_realm_default_subscriptions=False
+        )
+        self.check_user_able_to_register(email4, invite_link)
+        self.check_user_subscribed_only_to_streams(name4, [])
+
+        default_streams = get_slim_realm_default_streams(self.realm.id)
+        self.assert_length(default_streams, 3)
+        rome = get_stream("Rome", self.realm)
+        invite_link = self.generate_multiuse_invite_link(
+            streams=[rome], include_realm_default_subscriptions=True
+        )
+        self.check_user_able_to_register(email5, invite_link)
+        rome = get_stream("Rome", self.realm)
+        self.check_user_subscribed_only_to_streams(
+            name5, [rome, default_streams[0], default_streams[1], default_streams[2]]
+        )
 
     def test_multiuse_link_different_realms(self) -> None:
         """
@@ -2490,11 +2603,54 @@ class MultiuseInviteTest(ZulipTestCase):
 
         self.login("iago")
         stream_ids = []
+        default_streams = get_slim_realm_default_streams(self.realm.id)
+        verona = get_stream("Verona", self.realm)
+        sandbox = get_stream("sandbox", self.realm)
+        zulip = get_stream("Zulip", self.realm)
+        self.assertCountEqual(default_streams, [verona, sandbox, zulip])
+
+        # Check that user is subscribed to the streams that were set as default
+        # at the time of account creation and not at the time of inviting them.
         result = self.client_post(
             "/json/invites/multiuse",
             {
                 "stream_ids": orjson.dumps(stream_ids).decode(),
                 "invite_expires_in_minutes": 2 * 24 * 60,
+                "include_realm_default_subscriptions": orjson.dumps(True).decode(),
+            },
+        )
+        invite_link = self.assert_json_success(result)["invite_link"]
+
+        denmark = get_stream("Denmark", self.realm)
+        do_add_default_stream(denmark)
+        do_remove_default_stream(verona)
+        self.check_user_able_to_register(self.nonreg_email("test1"), invite_link)
+        self.check_user_subscribed_only_to_streams("test1", [denmark, sandbox, zulip])
+
+        stream_ids = [verona.id]
+        self.login("iago")
+        result = self.client_post(
+            "/json/invites/multiuse",
+            {
+                "stream_ids": orjson.dumps(stream_ids).decode(),
+                "invite_expires_in_minutes": 2 * 24 * 60,
+                "include_realm_default_subscriptions": orjson.dumps(True).decode(),
+            },
+        )
+        invite_link = self.assert_json_success(result)["invite_link"]
+        self.check_user_able_to_register(self.nonreg_email("newguy"), invite_link)
+        default_streams = get_slim_realm_default_streams(self.realm.id)
+        self.assertCountEqual(default_streams, [denmark, sandbox, zulip])
+        self.check_user_subscribed_only_to_streams("newguy", [denmark, sandbox, verona, zulip])
+
+        self.login("iago")
+        stream_ids = []
+        result = self.client_post(
+            "/json/invites/multiuse",
+            {
+                "stream_ids": orjson.dumps(stream_ids).decode(),
+                "invite_expires_in_minutes": 2 * 24 * 60,
+                "include_realm_default_subscriptions": orjson.dumps(False).decode(),
             },
         )
         invite_link = self.assert_json_success(result)["invite_link"]
@@ -2535,6 +2691,7 @@ class MultiuseInviteTest(ZulipTestCase):
             {
                 "stream_ids": orjson.dumps([]).decode(),
                 "invite_expires_in_minutes": 2 * 24 * 60,
+                "include_realm_default_subscriptions": orjson.dumps(True).decode(),
             },
         )
         invite_link = self.assert_json_success(result)["invite_link"]